+++ editor/changelog.md
@@ -3,6 +3,7 @@
 ## Solarus Editor 2.0.3 (In progress)
 
 * Fix crash on editor startup if the quest was not found (#1962).
+* Meaningful error message when trying to create a map when there is no tileset (#1954).
 
 ## Solarus Editor 2.0.2 (2025-12-05)